Add type annotations to model metadata
model.py has gotten pretty big, so split out Metadata into a separate
file and add type annotations to it. Also the Metadata got confusing
with admin SpannerMetadata, so I'm starting to rename things to make
more sense.
